requirements.txt
streamlit
rembg
Pillow
py2app
tqdm